+++ NEWS.md
@@ -1,5 +1,10 @@
 # mlr3spatial 0.1.2.9000
 
+* BREAKING CHANGE: `TaskClassifST` and `TaskRegrST` are used to train a learner with spatial data.
+* BREAKING CHANGE: Raster objects cannot be used to create tasks for training anymore.
+* BREAKING CHANGE: `TaskUnsupervised` is used to predict a raster now.
+* feat: Add `as_task_regr_st()` and `as_task_classif_st()` from spatial objects.
+* feat: Add `as_task_unsupervised()` from raster objects.
 * feat: `data("leipzig")` loads an `sf` object with land cover in Leipzig.
 * feat: GeoTIFF and GeoPackage of Leipzig in `extdata` folder.
 * feat: Task `leipzig` with land cover target.
@@ -7,7 +12,6 @@
 * BREAKING CHANGE: `DataBackendRaster` cannot be created from `RasterLayer` objects anymore.
 * fix: `DataBackendVector` initialization failed when geometry column had a different name than `geometry`.
 * fix: `spatial_predict()` returned an unnamed response.
-* feat: Add `as_task_regr()` and `as_task_classif()` from spatial objects.
 * fix: `spatial_predict()` wrote predictions to the wrong cell.
 * BREAKING CHANGE: Remove `demo_raster()`, `demo_stack_spatraster()`, `demo_stack_rasterbrick()` and `demo_rasterbrick()` functions.
 * feat: Prediction layer contains `NA` at raster cells with `NA` values in one or more feature layers.